requirements.txt
# =============================================================================
# Zero-Omission-Harness Dependencies
# =============================================================================

# Core dependencies
pyyaml>=6.0
jsonschema>=4.0.0
pathspec>=0.11.0

# CLI (optional but recommended)
typer>=0.9.0
rich>=13.0.0
click>=8.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: File watching for auto-validation
# watchdog>=3.0.0

# Optional: Git integration
# gitpython>=3.1.0
